+++ client/search-ui/src/components/CopyPathAction.tsx
@@ -1,28 +1,29 @@
 import React, { useState } from 'react'
 
 import { mdiContentCopy } from '@mdi/js'
+import classNames from 'classnames'
 import copy from 'copy-to-clipboard'
-import { useLocation } from 'react-router'
 
+import { TelemetryProps } from '@sourcegraph/shared/src/telemetry/telemetryService'
 import { Button, Icon, screenReaderAnnounce, Tooltip } from '@sourcegraph/wildcard'
-
-import { eventLogger } from '../../tracking/eventLogger'
-import { parseBrowserRepoURL } from '../../util/url'
 
 import styles from './CopyPathAction.module.scss'
 
 /**
  * A repository header action that copies the current page's repository or file path to the clipboard.
  */
-export const CopyPathAction: React.FunctionComponent<React.PropsWithChildren<unknown>> = () => {
-    const location = useLocation()
+export const CopyPathAction: React.FunctionComponent<
+    {
+        filePath: string
+        className?: string
+    } & TelemetryProps
+> = ({ filePath, className, telemetryService }) => {
     const [copied, setCopied] = useState(false)
 
     const onClick = (event: React.MouseEvent<HTMLButtonElement>): void => {
         event.preventDefault()
-        eventLogger.log('CopyFilePath')
-        const { repoName, filePath } = parseBrowserRepoURL(location.pathname)
-        copy(filePath || repoName) // copy the file path if present; else it's the repo path.
+        telemetryService.log('CopyFilePath')
+        copy(filePath)
         setCopied(true)
         screenReaderAnnounce('Path copied to clipboard')
 
@@ -35,7 +36,13 @@ export const CopyPathAction: React.FunctionComponent<React.PropsWithChildren<unk
 
     return (
         <Tooltip content={label}>
-            <Button aria-label="Copy" variant="icon" className="p-2" onClick={onClick} size="sm">
+            <Button
+                aria-label="Copy path to clipboard"
+                variant="icon"
+                className={classNames(styles.button, className)}
+                onClick={onClick}
+                size="sm"
+            >
                 <Icon className={styles.copyIcon} aria-hidden={true} svgPath={mdiContentCopy} />
             </Button>
         </Tooltip>