+++ client/shared/src/search/query/hover.ts
@@ -172,6 +172,10 @@ const toPredicateHover = (token: MetaPredicate): string => {
             return `**Built-in predicate**. Search only inside repositories that have been committed to since \`${parameters}\`.`
         case 'has.description':
             return '**Built-in predicate**. Search only inside repositories that have a **description** matching the given regular expression'
+        case 'has.tag':
+            return '**Built-in predicate**. Search only inside repositories that are tagged with the given tag'
+        case 'has':
+            return '**Built-in predicate**. Search only inside repositories that are associated with the given key:value pair'
     }
     return ''
 }