+++ client/web/src/components/fuzzyFinder/FuzzyFinder.tsx
@@ -1,78 +1,91 @@
-import React, { useState, useEffect, Dispatch, SetStateAction } from 'react'
+import React, { useEffect, Dispatch, SetStateAction, useCallback } from 'react'
 
-import { ApolloError, useQuery } from '@apollo/client'
 import * as H from 'history'
-import { useHistory } from 'react-router-dom'
 
-import { gql, getDocumentNode } from '@sourcegraph/http-client'
-import { useKeyboardShortcut } from '@sourcegraph/shared/src/keyboardShortcuts/useKeyboardShortcut'
 import { Shortcut } from '@sourcegraph/shared/src/react-shortcuts'
+import { SettingsCascadeProps } from '@sourcegraph/shared/src/settings/settings'
 import { TelemetryProps } from '@sourcegraph/shared/src/telemetry/telemetryService'
 
-import { FuzzySearch, SearchIndexing } from '../../fuzzyFinder/FuzzySearch'
-import { FileNamesResult, FileNamesVariables } from '../../graphql-operations'
-import { parseBrowserRepoURL } from '../../util/url'
-
 import { FuzzyModal } from './FuzzyModal'
+import { useFuzzyShortcuts } from './FuzzyShortcuts'
+import { fuzzyIsActive, FuzzyTabsProps, FuzzyState, useFuzzyState, FuzzyTabKey } from './FuzzyTabs'
 
 const DEFAULT_MAX_RESULTS = 100
 
-interface FuzzyFinderContainerProps extends TelemetryProps, Pick<FuzzyFinderProps, 'location'> {}
+export interface FuzzyFinderContainerProps
+    extends TelemetryProps,
+        Pick<FuzzyFinderProps, 'location'>,
+        SettingsCascadeProps,
+        FuzzyTabsProps {
+    isVisible: boolean
+    setIsVisible: React.Dispatch<SetStateAction<boolean>>
+}
 
 /**
  * This components registers a global keyboard shortcut to render the fuzzy
  * finder and renders the fuzzy finder.
  */
-export const FuzzyFinderContainer: React.FunctionComponent<FuzzyFinderContainerProps> = ({
-    location,
-    telemetryService,
-}) => {
-    const [isVisible, setIsVisible] = useState(false)
-    const [retainFuzzyFinderCache, setRetainFuzzyFinderCache] = useState(true)
-    const fuzzyFinderShortcut = useKeyboardShortcut('fuzzyFinder')
+export const FuzzyFinderContainer: React.FunctionComponent<FuzzyFinderContainerProps> = props => {
+    const { isVisible, setIsVisible } = props
+    const state = useFuzzyState(props, () => setIsVisible(false))
+    const { tabs, activeTab, setActiveTab, repoRevision } = state
+    const shortcuts = useFuzzyShortcuts(props.settingsCascade.final)
+
+    const openFuzzyFinder = useCallback(
+        (tab: FuzzyTabKey): void => {
+            const newTab = tabs.focusNamedTab(tab)
+            if (newTab) {
+                setActiveTab(newTab)
+            }
+            if (isVisible) {
+                return
+            }
+            setIsVisible(true)
+        },
+        [tabs, setActiveTab, isVisible, setIsVisible]
+    )
 
     useEffect(() => {
         if (isVisible) {
-            telemetryService.log('FuzzyFinderViewed', { action: 'shortcut open' })
+            props.telemetryService.log('FuzzyFinderViewed', { action: 'shortcut open' })
         }
-    }, [telemetryService, isVisible])
+    }, [props.telemetryService, isVisible])
+
+    if (tabs.isAllDisabled()) {
+        return null
+    }
+
+    // Disable the fuzzy finder if only the 'files' tab is enabled and we're not
+    // in a repository-related page.
+    if (tabs.isOnlyFilesEnabled() && !fuzzyIsActive(activeTab, repoRevision, 'files')) {
+        return null
+    }
 
     return (
         <>
-            {fuzzyFinderShortcut?.keybindings.map((keybinding, index) => (
-                <Shortcut
-                    key={index}
-                    {...keybinding}
-                    onMatch={() => {
-                        setIsVisible(true)
-                        setRetainFuzzyFinderCache(true)
-                        const input = document.querySelector<HTMLInputElement>('#fuzzy-modal-input')
-                        input?.focus()
-                        input?.select()
-                    }}
-                    ignoreInput={true}
-                />
-            ))}
-            {retainFuzzyFinderCache && (
-                <FuzzyFinder
-                    setIsVisible={bool => setIsVisible(bool)}
-                    isVisible={isVisible}
-                    location={location}
-                    setCacheRetention={bool => setRetainFuzzyFinderCache(bool)}
-                />
+            {shortcuts
+                .filter(shortcut => shortcut.isEnabled)
+                .flatMap(shortcut =>
+                    shortcut.shortcut?.keybindings.map(keybinding => (
+                        <Shortcut
+                            {...keybinding}
+                            key={`fuzzy-shortcut-${shortcut.name}-${JSON.stringify(keybinding)}`}
+                            onMatch={() => openFuzzyFinder(shortcut.name)}
+                            ignoreInput={true}
+                        />
+                    ))
+                )}
+            {isVisible && (
+                <FuzzyFinder {...state} setIsVisible={bool => setIsVisible(bool)} location={props.location} />
             )}
         </>
     )
 }
 
-interface FuzzyFinderProps {
+interface FuzzyFinderProps extends FuzzyState {
     setIsVisible: Dispatch<SetStateAction<boolean>>
 
-    isVisible: boolean
-
     location: H.Location
-
-    setCacheRetention: Dispatch<SetStateAction<boolean>>
 
     /**
      * The maximum number of files a repo can have to use case-insensitive fuzzy finding.
@@ -85,119 +98,9 @@ interface FuzzyFinderProps {
     caseInsensitiveFileCountThreshold?: number
 }
 
-const FuzzyFinder: React.FunctionComponent<React.PropsWithChildren<FuzzyFinderProps>> = ({
-    location: { search, pathname, hash },
-    setCacheRetention,
-    setIsVisible,
-    isVisible,
-}) => {
-    // The state machine of the fuzzy finder. See `FuzzyFSM` for more details
-    // about the state transititions.
-    const [fsm, setFsm] = useState<FuzzyFSM>({ key: 'empty' })
-    const { repoName = '', commitID = '', rawRevision = '' } = parseBrowserRepoURL(pathname + search + hash)
-    const { downloadFilename, isLoadingFilename, filenameError } = useFilename(repoName, commitID || rawRevision)
+const FuzzyFinder: React.FunctionComponent<React.PropsWithChildren<FuzzyFinderProps>> = props => {
+    const { setIsVisible } = props
+    const onClose = useCallback(() => setIsVisible(false), [setIsVisible])
 
-    const history = useHistory()
-    useEffect(
-        () =>
-            history.listen(location => {
-                const url = location.pathname + location.search + location.hash
-                const { repoName: repo = '', commitID: commit = '', rawRevision: raw = '' } = parseBrowserRepoURL(url)
-                if (repo !== repoName || commit !== commitID || raw !== rawRevision) {
-                    setCacheRetention(false)
-                }
-            }),
-        [history, repoName, commitID, rawRevision, setCacheRetention]
-    )
-
-    if (!isVisible) {
-        return null
-    }
-
-    return (
-        <FuzzyModal
-            repoName={repoName}
-            commitID={commitID}
-            initialMaxResults={DEFAULT_MAX_RESULTS}
-            initialQuery=""
-            downloadFilenames={downloadFilename}
-            isLoading={isLoadingFilename}
-            isError={filenameError}
-            onClose={() => setIsVisible(false)}
-            fsm={fsm}
-            setFsm={setFsm}
-        />
-    )
-}
-
-/**
- * The fuzzy finder modal is implemented as a state machine with the following transitions:
- *
- * ```
- *   ╭────[cached]───────────────────────╮  ╭──╮
- *   │                                   v  │  v
- * Empty ─[uncached]───> Downloading ──> Indexing ──> Ready
- *                       ╰──────────────────────> Failed
- * ```
- *
- * - Empty: start state.
- * - Downloading: downloading filenames from the remote server. The filenames
- *                are cached using the browser's CacheStorage, if available.
- * - Indexing: processing the downloaded filenames. This step is usually
- *             instant, unless the repo is very large (>100k source files).
- *             In the torvalds/linux repo (~70k files), this step takes <1s
- *             on my computer but the chromium/chromium repo (~360k files)
- *             it takes ~3-5 seconds. This step is async so that the user can
- *             query against partially indexed results.
- * - Ready: all filenames have been indexed.
- * - Failed: something unexpected happened, the user can't fuzzy find files.
- */
-export type FuzzyFSM = Empty | Downloading | Indexing | Ready | Failed
-export interface Empty {
-    key: 'empty'
-}
-export interface Downloading {
-    key: 'downloading'
-}
-export interface Indexing {
-    key: 'indexing'
-    indexing: SearchIndexing
-}
-export interface Ready {
-    key: 'ready'
-    fuzzy: FuzzySearch
-}
-export interface Failed {
-    key: 'failed'
-    errorMessage: string
-}
-
-const FILE_NAMES = gql`
-    query FileNames($repository: String!, $commit: String!) {
-        repository(name: $repository) {
-            id
-            commit(rev: $commit) {
-                id
-                fileNames
-            }
-        }
-    }
-`
-
-interface FilenameResult {
-    downloadFilename: string[]
-    isLoadingFilename: boolean
-    filenameError: ApolloError | undefined
-}
-
-const useFilename = (repository: string, commit: string): FilenameResult => {
-    const { data, loading, error } = useQuery<FileNamesResult, FileNamesVariables>(getDocumentNode(FILE_NAMES), {
-        variables: { repository, commit },
-    })
-
-    return {
-        downloadFilename: data?.repository?.commit?.fileNames || [],
-        isLoadingFilename: loading,
-        filenameError: error,
-    }
+    return <FuzzyModal {...props} initialMaxResults={DEFAULT_MAX_RESULTS} initialQuery="" onClose={onClose} />
 }