+++ cmd/frontend/internal/cli/serve_cmd.go
@@ -122,7 +122,7 @@ func InitDB(logger sglog.Logger) (*sql.DB, error) {
 }
 
 // Main is the main entrypoint for the frontend server program.
-func Main(enterpriseSetupHook func(db database.DB, c conftypes.UnifiedWatchable) enterprise.Services) error {
+func Main(enterpriseSetupHook func(db database.DB, codeIntelServices codeintel.Services, c conftypes.UnifiedWatchable) enterprise.Services) error {
 	ctx := context.Background()
 
 	log.SetFlags(0)
@@ -177,6 +177,15 @@ func Main(enterpriseSetupHook func(db database.DB, c conftypes.UnifiedWatchable)
 	conf.MustValidateDefaults()
 	go conf.Watch(liblog.Update(conf.GetLogSinks))
 
+	codeIntelServices, err := codeintel.GetServices(codeintel.Databases{
+		DB: db,
+		// N.B. must call after conf.Init()
+		CodeIntelDB: mustInitializeCodeIntelDB(logger),
+	})
+	if err != nil {
+		return err
+	}
+
 	// now we can init the keyring, as it depends on site config
 	if err := keyring.Init(ctx); err != nil {
 		return errors.Wrap(err, "failed to initialize encryption keyring")
@@ -199,13 +208,13 @@ func Main(enterpriseSetupHook func(db database.DB, c conftypes.UnifiedWatchable)
 	profiler.Init()
 
 	// Run enterprise setup hook
-	enterprise := enterpriseSetupHook(db, conf.DefaultClient())
+	enterprise := enterpriseSetupHook(db, codeIntelServices, conf.DefaultClient())
 
 	authz.DefaultSubRepoPermsChecker, err = authz.NewSubRepoPermsClient(db.SubRepoPerms())
 	if err != nil {
 		return errors.Wrap(err, "Failed to create sub-repo client")
 	}
-	ui.InitRouter(db, enterprise.CodeIntelResolver)
+	ui.InitRouter(db)
 
 	if len(os.Args) >= 2 {
 		switch os.Args[1] {
@@ -256,6 +265,7 @@ func Main(enterpriseSetupHook func(db database.DB, c conftypes.UnifiedWatchable)
 
 	siteid.Init(db)
 
+	globals.WatchBranding()
 	globals.WatchExternalURL(defaultExternalURL(nginxAddr, httpAddr))
 	globals.WatchPermissionsUserMapping()
 
@@ -291,16 +301,6 @@ func Main(enterpriseSetupHook func(db database.DB, c conftypes.UnifiedWatchable)
 	}
 
 	server, err := makeExternalAPI(db, schema, enterprise, rateLimitWatcher)
-	if err != nil {
-		return err
-	}
-
-	codeIntelDB := mustInitializeCodeIntelDB(logger)
-
-	codeIntelServices, err := codeintel.GetServices(codeintel.Databases{
-		DB:          db,
-		CodeIntelDB: codeIntelDB,
-	})
 	if err != nil {
 		return err
 	}
@@ -439,9 +439,11 @@ func mustInitializeCodeIntelDB(logger sglog.Logger) stores.CodeIntelDB {
 	dsn := conf.GetServiceConnectionValueAndRestartOnChange(func(serviceConnections conftypes.ServiceConnections) string {
 		return serviceConnections.CodeIntelPostgresDSN
 	})
-	db, err := connections.EnsureNewCodeIntelDB(dsn, "frontend-internal", &observation.TestContext)
+
+	db, err := connections.EnsureNewCodeIntelDB(dsn, "frontend", &observation.TestContext)
 	if err != nil {
 		logger.Fatal("Failed to connect to codeintel database", sglog.Error(err))
 	}
+
 	return stores.NewCodeIntelDB(db)
 }